+++ Shuhari.Framework.Common/ComponentModel/ISelectable.cs
@@ -1,0 +1,13 @@
+﻿namespace Shuhari.Framework.ComponentModel
+{
+    /// <summary>
+    /// Selected elemnent
+    /// </summary>
+    public interface ISelectable
+    {
+        /// <summary>
+        /// If selected
+        /// </summary>
+        bool Selected { get; set; }
+    }
+}