+++ Shuhari.Framework.Wpf/Models/LogType.cs
@@ -1,0 +1,28 @@
+﻿namespace Shuhari.Framework.Wpf.Models
+{
+    /// <summary>
+    /// Log Type
+    /// </summary>
+    public enum LogType
+    {
+        /// <summary>
+        /// Unknown
+        /// </summary>
+        Unknown = 0,
+
+        /// <summary>
+        /// Info
+        /// </summary>
+        Info = 1,
+
+        /// <summary>
+        /// Warning
+        /// </summary>
+        Warn = 2,
+
+        /// <summary>
+        /// Error 
+        /// </summary>
+        Error = 3,
+    }
+}